feat: add notification

+++ payment-service/src/main/java/com/lumina_bank/paymentservice/domain/repository/ExchangeRateRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ExchangeRateRepository extends JpaRepository<ExchangeRate, Integer> {
+public interface ExchangeRateRepository extends JpaRepository<ExchangeRate, Long> {
     Optional<ExchangeRate> findTopByBaseCurrencyAndTargetCurrencyOrderByDateDesc(Currency from, Currency to);
 }
